Move function to `esphome.config_helpers`
